feat: Add Permit.io configuration to settings.py
This commit adds the necessary configuration for Permit.io to the `settings.py` file in the `hsystem` project. It includes the addition of the `PERMIT_API_KEY` and `PERMIT_PDP_URL` environment variables, which are loaded from the `.env` file. This allows developers to easily configure Permit.io by updating the values in the `.env` file.